Use ujson instead of simplejson
This makes a big difference. About 81% of the WSGI app's time is currently
spent in calls to simplejson.dumps(), which itself takes about 50% more time
than just using 'json' from the standard library under CPython 2.7.

Since simplejson is an old third-party library that relies on C extensions to
get speedups under CPython, it seems fair to replace it with a newer
third-party library, otherwise we are crippling what should be Python's
performance baseline.