Generate configuration cache before starting benchmarks
Symfony relies on configuration caching to avoid booting the whole system on every request over and over again. If you are not doing this during the build process, then the application might run into cache slams during the first 2-3 second when the cache is generated the first time.

Compared to lightweight frameworks such as Laravel or Silex, Symfony uses lots of configuration files that are parsed and compiled into two php files (one for the services, one for the routing). This is comparable to the compile step of any static language and only done once. It is not a runtime cache.