importing the end product, dumps
And omitting another dot-notation __getattr__ call.  This technique can get a little ridiculous as it's hard to tell which modules the functions came from and face it, uj_dumps is a name I came up with, not a standard name in the API for the module.
On the JSON response test, it will possibly give detectable improvements.