Heavily stripped down version of aspnet tests: twice as fast for some tests
Based on pdonald's aspnet tests, this is a heavily stripped down version
that pulls every trick in the book, to double the performance of CPU bound
tests. Perhaps this is not production worthy, but should provide a good
basis of comparison against microframeworks/microplatforms that barely do
anything.

See README.md for more info.