Get aspnet-stripped working with MySQL, PostgreSQL, MongoDB
Since @pdonald did all the work to get ASP.NET working against these other
databases, we should get this working in aspnet-stripped, so that we can
see how much stripping improved performance with these databases. This
ports @pdonald's related changes from aspnet to aspnet-stripped.

Because some of the URLs have the '&' character in them, I had to fix
framework_test.py to do proper quoting so that the '&' character would not
be interpreted by the shell.