uwsgi: use multiple http router processes for 30% performance gain
When using a single http router process (the default when --http is used),
all CPU cores were not saturated. I think this is because the single http
router process became CPU bound and the bottleneck. Thus, use
--http-processes PROCS to have one router process per core. This enables
all cores to be saturated and improved throughput by 30% on my setup.

I learned about --http-processes from the creator of uWSGI on the uWSGI
mailing list:

http://lists.unbit.it/pipermail/uwsgi/2013-September/006422.html