18938 Break logging into individual files
Previously, all "logging" was sent to standard out, and we would
capture that in a file. The problem with this is that all of the
frameworks would be logged to the same file, resulting in an
extremely large log.

This change causes each individual framework to create a log file
in the results folder, and captures the standard output there.
Additionally, each setup file now has start/stop functions that
require a 'logfile' argument. Initially, basically all of the tests
do not use this file, but this gives the possibility to attach
the logfile as the capture for stderr/stdout on the subprocesses
spawned when a server is spun up.

Similarly, 'logfile.write()' will allow a developer to enter logs
directly for debug or general purposes.