verify Huge change to parsing and validation
Before this commit, if the test was marked 'json', then a raw file
would be created for that test in the
`results/[timestamp]/json/[your test]/`

However, if a test was marked 'json' and it failed, the raw file
would not be output to the
`results/[timestamp]/json/[your test]/`
dir which would mean there was no way to tell, after a run, whether
the test passed or failed validation.

With this commit, if your test is marked 'json', and it succeeds,
nothing has changed - your test gets a
`results/[timestamp]/json/[your test]/raw`
file with your timing data in it.

If your test is NOT marked 'json', then you will not have a
`results/[timestamp]/json/[your test]/raw`
file. This means you did not 'fail' your json test, but simply did
not have a json test.

If your test is marked 'json' and it failed validation, then your
`results/[timestamp]/json/[your test]/raw`
file will exist, but will be empty. This file existing and being
empty will determine at parse-time whether your test failed
validation or not.