verify Raw file was not being created correctly
Each test's `raw` file was only being created if the test was
successfully benchmarked, but we need the file to exist if the
test was only MARKED as having should benchmark.

Confusing...

Okay, so if your test has a `json` test specified, but it did not
pass validation, we skip benchmarking it (as there is no good reason
to benchmark a test that is broken/cheating/improper/etc). The
problem is that in skipping the test, we never created the `raw`
file in the `results/[name]/[timestamp]/json/[test name]/` dir,
which means that when we parse the results, instead of appending
`[test name]` to the list of failed, we ignored it.

This commit touches the `raw` file in the benchmarking step
regardless of whether the validation was successful. We still skip
benchmarking if the test was not valid, but now the `raw` file for
testX's `json` will exist and be empty, giving us enough information
to suggest it failed during parsing.