verify Fixed some type errors and reduced code
Basically, some tests were returning strings for the values
of 'id' and 'randomNumber', which is not necessarily wrong
since JS can deal with either equally well through double-
equals style equality.

Other tests were returning floats instead of integers, but
again these are fine because JS handles all number types as
floats anyway.

What I ultimately did was change how we validate number-
expected values. Now, I parse them as floats inside of a try/
catch block; this will successfully convert 123, "123",
"123.0" and 123.0 to 123.0, but fail on non-numerical values.
After parsing to float, I check that each type is float and
return true if so... this is mostly because I did not know
the best practices in Python for having unassigned values
and using try/catch blocks for control flow. Likely, this is
a frowned upon practice...