verify Fixed a bug with white space
It seemed that for certain tests who escaped the dash character
an error was occuring. Essentially, the data would be read until
the point of the `&mdash;` and then appended to our normalized
test string after being stripped, which means that the space
char before and after the dash char would get stripped and the
normalized string would not be correct because a few spaces
would be missing.

Ultimately, I am fine with having the data NOT stripped of all
whitespace since it will exist between begin and end tags and
every test should strive for perfection; a leading/trailing
whitespace character should result in a failure.