verify Allow apostrophes for now
This is a touchy subject, but there are so many frameworks
that use third-party libraries for escaping html entities
and not all of them agree on apostrophes. To ease the
transition, I am going to simply allow apostrophes FOR NOW.

I may (read: likely will) make escaping apostrophes in the
fortunes test mandatory in the future.