verify Changed to explicitly require </html>
This was added originally because of a single framework that
seemed to be omitting </html>, which would result in the same
page rendering and was technically valid html5.

This check was readded because we are striving to maintain an
"apples-to-apples" response body. My worry is that if we do
not strictly require </html>, then some performant frameworks
might end up "optimizing" their fortunes test to omit it.