PR738 Not sure why, but this needed to be remade
Seems the python json parser was not happy with the invisibile
change made to the first line's opening object brace. I have no
idea why, but when I replaced the file with the raw contents of
the file and saved, something was changed and it seemed to fix
the problem.