remove spray for now
Two reasons:

1) The tested version of spray is quite old by now and several things have
changed.
2) There seems to be something wrong with spray on the new hardware [1]
but we currently have no capacity to update spray's benchmark or debug
the issues with the new hardware.

The conclusion is that neither the version nor the results are currently
representative of spray's performance.

We are currently working on a version (which will be akka-http)
which is not yet ready. We will offer a new version of this benchmark
once the new version is ready.

[1] https://groups.google.com/d/msg/framework-benchmarks/0XrAVkcjmT0/ChyyA3V8h70J