Corrects ERR trap
function fw_exists was returning a non-zero code, which was always
triggering the ERR trap. This commit fixes that issue for the "go"
dependency - it still needs to be fixed for others.

This also updates the ERR trap to have access to the entire
bash function stack, which would be very useful for debugging