Duda I/O: adding Duda I/O framework tests.
The following patch adds a web service based on Duda I/O framework,
at the moment only tests #1 and #6 are implemented. For the next round
we will cover Database tests.

For more details about the stack and components please refer to:

http://duda.io
http://monkey-project.com

note: this is the second pull request upon requirement to drop the stack
from the patch. This PR replaces #905.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>