FixRunTests fixes a bug with truthy strings
Essentially, the bug was that the config reader would read in
the benchmark.cfg file as a string->string dictionary, and then
we would test the truthiness of the values. This ends up being
basically a `if bool("False"):` which evaluates to True.

Nothing can be set to `False` in the benchmark.cfg, so I ended
up commenting out those lines and setting their values to True
instead. This will imply that they are defaulted to False and
you should only uncomment them if you need to set them to True.