Updated how we invoke the start scripts
Before we were calling Popen which spawns a new process
and continues in a non-blocking way. This can (and did)
lead to some longer-running setup scripts to not be ready
when we tried to verify. Additionally, it was hard to
determine the error in the case where the setup.sh file
failed to run successfully.

Therefore, I can changed the call back to a blocking one
(check_call) and added back the retcode and logging in
the case of errors. I tested this and if you return a
non-zero result, you get some nice logging, and if you
return nothing it continues as expected.

One caveat to this is that the startup scripts needed a
way to run a given script in a non-blocking way. For
example, when frameworkX starts the server, it just sits
and waits which would never return. Therefore, any/all
tests that perform this way should invoke their server
start code in the background, which will not block.

Example:
java myserver.jar &

That will start a process in the background as the current
user and move on.