Fixed an annoying bug with home dir
Essentially, using `sudo -u [user] -E` caused the full
environment to be copied over, including the home dir
of the current user. The problem is that `[user]` cannot
write to the home directory of the launching user.

Adding the -H flag causes the home directory to be
forcibly reset to the user specified with -u.