Attempts to convert setup.py/setup-raw.py to setup.sh/setup-raw.sh for php-senthot and php-silex.
Was unable to get php-senthot to run test, confirmed that it did not complete
in round 9 testing. Left the setup.py file intact in the event further
research will be done.

php-silex completed with warnings, but php-silex-raw failed all tests but
json. Confirmed that Round 9 testing did not complete. Removed .py files
as the tests ran, just failed.