Fix incorrect results.
Previously the play-scala-mongodb test handled both the single query and
multiple query tests through /db (maybe this was the format of an older
round of TFB?), and so returned a list for the single query test and
couldn't handle non-integer or out-of-range requests for the multi
query test.  Now both should pass.