Add Cutelyst framework tests
This includes all 6 test as
well as json configuration for
running them. Since Cutelyst has
a pluggable Engine which the best one is
uwsgi, the test use it to process the
HTTP requests directly, an initial work
to use nginx as a front-end giving it
keep-alive support is also available.

The Qt 5.6 PPA used for these test did
not provide a postgres driver so that's
disabled at the moment.