45216 - Adds continuous benchmarking to the TFB project.
toolset/run-continuously.sh is simply a while loop that
removes and rebuilds the framework benchmark suite and
then runs the benchmarks.

There are five environment variables to be set:

* TFB_REPOPARENT     the absolute path to the folder containing
  the repository
* TFB_REPONAME       the name of the repository folder
* TFB_REPOURI        the URI of the repository
* TFB_REPOBRANCH     the branch of the repository to use
* TFB_LOGSFOLDER     the folder to archive log files

There are four life cycle stages in run-continuously:

* tear-down-environment.sh  called upon to remove the existing
  environment
* rebuild-environment.sh    called upon to rebuild the environment
* pre-run-tests             scripts in this folder are called
                            before the benchmarks are run
* post-run-tests            scripts in this folder are called
                            after the benchmarks are run

- Notes -

run-continuously generally assumes that a clone of the
appropriate repo and branch exist and have the list
scripts available and an appropriate copy of benchmark.cfg
in place.  Some effort is made to support starting states
that differ but those are not intended to be the general
case for use.