H2O: Optimize the multiple queries test
Previously, the individual database queries that were used to
generate the response were executed serially in order to simplify
the implementation (error handling in particular). The underlying
assumption was that there would be enough parallel HTTP requests
(each resulting in a single query at a time) to saturate the
database connection pool. However, the multiple queries test is
performed with a fixed number of concurrent HTTP connections (256),
and the current default number of database connections per thread
is 5, so that assumption is no longer true when the machine has
more than 25 cores (keeping in mind that standard server processors
implement simultaneous multithreading, e.g. Hyper-threading, and
that the program starts a thread per core by default) without even
considering any issues with the load balancing among the threads.

This commit changes the individual queries to be executed in
parallel (with the concurrency limited by the number of database
connections per thread). Furthermore, the updates test will be
disabled until it receives the same treatment.