Had to daemonize TFBReaper
The old implementation had a race-condition with wait()
that would sometimes (often?) result in the first process
to exit successfully and fire off the wait() function
leaving the child processes orphaned as TFBReaper exited.

The new implementation drops wait() in favor of simply
entering an infinite loop and relying on the suite to handle
the process cleanup.

This new implementation has the added wrinkle of TFBReaper
remaining as a 'defunct' process because it had been exited
forcibly and gone into a state of waiting for its parent
(TFB python suite) to clean it or exit. This would result
in hundreds of defunct TFBReaper processes left running as
a full benchmark neared conclusion (each test would spawn
a new-and-eventually-defunct TFBReaper). The fix to this
issue is actually the original problem - have TFBReaper
fork a child process and exit the parent.

This causes TFBReaper's child process to become orphaned
and adopted by init(1), which will clean defunct processes
by design almost immediately.