Added a SIGSTOP to try and end forking
Essentially, a SIGTERM to the wrong process before
the right process could result in a fork() that causes
a rogue process to not be SIGTERMed. The SIGKILL ends
up cleaning this out anyway, but we really want to
avoid that if possible.

Sending SIGSTOP to all the processes before SIGTERM
tells them to stop allowing fork() calls. Then, when
we call SIGTERM we should not have any rogue processes
as the result of unprotected fork()ing.

Also, dropped some unneeded sleep() calls that just
slowed down each test run.