H2O: Optimize the database updates test
The database updates test receives the same treatment as the multiple
queries test before. Furthermore, the World objects were previously
persisted with their new values at the end of the process (using batch
updates) after reading the database completed, but now each object is
read and updated in a single transaction, which reduces the roundtrips
to the database.