H2O: Make a couple of improvements
* Make the choice of the maximum number of database connections
smarter - the program does not have a global database connection
pool that is shared among all the threads; instead, there is
a separate connection pool per thread (since a connection pool
is not immutable state, and the policy has been to limit sharing
to such type of data as much as possible), and the maximum
number of connections applies to it individually. Furthermore,
there is an important difference between the cloud and the
physical hardware environments - while in the former the
application and the database server are the same, in the latter
the application server has more CPU cores. As a result, using
the same value for the maximum number of database connections
per thread in both environments means that in at least one of
them the setting will be suboptimal. That's why now the setup
script detects the current environment and adjusts the connection
number accordingly.
* Clean up the compiler flags - enabling the stack protector
by default in debug builds might interfere with other memory
debugging tools. Also, there is no need for the optimizations
enabled by "-Ofast".