Expand tokio-minihttp with /db and /queries (#2506)
This commit expands the Rust framework tokio-minihttp with routes for both `/db`
and `/queries`, implementing those two benchmarks. The Tokio framework in Rust
is asynchronous but there aren't many asynchronous database drivers just yet, so
these routes are implemented with blocking I/O on an external thread pool.

I also refactored the existing code a bit to be more amenable to the addition of
the new routes, and I also added a `Cargo.lock` file to ensure reproducible
builds in development and on the release servers. A `Cargo.lock` is similar to a
`Gemfile.lock` in Ruby which I've seen checked in for other frameworks in this
repo, so I figured it'd be a good idea to check it in here as well.