Avoid "sudo make install", which breaks some frameworks
In particular, using sudo here causes stuff in the "installs" directory
to be owned by root.  Then if another framework wanted to put something
in one of the same subdirectories of installs, and it was (properly) not
using sudo, it would fail.  For example, this happened with cutelyst
(used sudo inappropriately, created an installs/bin owned by root) and
ffead-cpp (correctly did *not* use sudo, could not add to installs/bin
because it was owned by root).

The fix for cmake is a bit messier.  cmake-3.5.sh was overriding whatever
version of cmake that other frameworks had installed, which was not good.
It turns out that swift/vapor run just fine using the same version of
cmake (3.2 or something) as the other frameworks, so I got rid of the
cmake-3.5.sh script and used apt-get instead.  I copied the related lines
from cutelyst -- another framework that uses cmake.

Ideally we should be using fw_depends instead of apt-get in the cutelyst
and swift installation scripts, but that was not the focus of this
change.  This change is about fixing file permissions.