Fix broken wicket db tests by downgrading the mysql connector
Version 6.x of the connector is apparently more opinionated about time
zones.  All of the db tests were failing with exceptions like this:

  java.sql.SQLException: The server time zone value 'CDT' is
  unrecognized or represents more than one time zone. You must
  configure either the server or JDBC driver (via the serverTimezone
  configuration property) to use a more specifc time zone value if you
  want to utilize time zone support.

Maybe it's making a valid point that we should address, maybe not.  But
right now mysql-connector-java 6.x is a "development release".  Let's
use the stable release (5.x) instead, which is what every other
framework that uses Java and mysql is doing already.