Repair lapis database tests by providing the correct db hostname
The lapis tests that required db access were all failing to connect to
the database.  This was caused by two issues:

  1. We were not passing along the "DBHOST" environment variable from
     nginx to lapis.  The fix was to explicitly share that variable in
     nginx.conf.

  2. The "DBHOST" environment variable (at least in our current
     dedicated hardware environment) is set to "TFB-database", a name
     defined in our /etc/hosts file.  lapis doesn't know about names
     defined in /etc/hosts.  The fix was to change the "DBHOST"
     environment variable to be the resolved address of what it was
     previously, so for example "TFB-database" would become the IP
     address of that server in our dedicated hardware environment.
     This *should* be compatible with every other use of the "DBHOST"
     environment variable...  It has always been the case that this
     variable could have been defined in the benchmark config as an
     IP address.

The change to config.moon is not strictly necessary and might not even
be syntactically valid.  The tests seem to ignore the config.moon file
and use the config.lua file instead.  I decided to make the change to
the .moon file anyway because it's closer to correct than what was
there before (at least it communicates the right intent).