Execute "revel run" in the foreground
The "revel run" command compiles the test code and launches the
application server.  If the test code has compilation errors, then the
command fails and does not launch the server.

If the command is run in the foreground, then a failure will be noticed
by our toolset and the test will terminate in a timely fashion.  If the
command is run in the background, however, then a failure will not be
noticed and our toolset will enter a loop, waiting for the application
server port to be bound.  In the failure case, the port is never bound
and the toolset ends up waiting for 105 minutes before finally giving
up.

In the success case, whether the application runs in the foreground or
background makes no difference.  It's ok for the setup script to never
terminate as long as it successfully launches the application server.
The toolset will eventually notice that it binds to the expected port
and will then run the appropriate benchmarks.

The 105 minute timeout seems extremely large, but the reason it's that
large is that there is at least one framework (wt) can take on the order
of one hour to compile.