Use mysql connector 5.x instead of 6.x in light-java-mysql (#2636)
Our ServerCentral dedicated hardware environment is set up in such a
way that breaks mysql connector 6.x with exceptions like this:

  java.sql.SQLException: The server time zone value 'PDT' is
  unrecognized or represents more than one time zone. You must
  configure either the server or JDBC driver (via the serverTimezone
  configuration property) to use a more specifc time zone value if you
  want to utilize time zone support.

As far as I know, 6.x is presently an unstable development release
whereas 5.x is the stable, production-ready release.  Other Java
frameworks in the benchmarks use 5.x.  It doesn't seem worthwhile to try
to fix this "problem" with our environment right now...  Maybe
later when 6.x is stable or we aren't trying to get a new round of
benchmarks out the door.