Make wicket encode its fortunes response in UTF-8 (#2664)
It appeared to be using the system default charset instead, which was
not necessarily UTF-8.

I removed the initialCapacity argument for the ArrayList because there's
no good reason to set it to 10,000.  Since the test requirements state
that the list shouldn't be sized using foreknowledge of the row count,
we might as well trust ArrayList's defaults.