H2O: Various improvements based on the preview data (#2673)
* Update H2O to version 2.2.0

* Enable link-time optimization for mustache-c

* H2O: Another attempt at batch updates

This commit changes the database updates test implementation
to perform all updates in a single batch again, hopefully in
a faster way than before.

* H2O: Reduce the number of threads in the physical hardware environment

Many of the tests in the physical hardware environment do not saturate
a gigabit network connection, so there is no need to start the same
number of threads as logical processors. In this way thread contention
(mainly in kernel space) is reduced, hence performance is improved.