H2O: Fix several minor issues (#2746)
* Reduce the peak memory usage caused by reusing JSON generators
in the multiple queries and database updates tests by not allocating
the generators in advance, but only when necessary.
* Make some additions to the server shutdown code that have been
missed when simultaneous support for HTTP and HTTPS has been
implemented.
* Set the Web server back in the framework metadata - the H2O project
the framework depends on provides two ways to use the Web server. One
of them is to run a standalone server (and use FastCGI, etc.), and
the other is to link to it as a library in the framework process.
Even though the implementation uses the second approach, IMHO setting
H2O as the Web server in the framework metadata is still the correct
thing to do because libh2o is used unmodified, and is pretty much
equivalent to running the standalone server (except for some details
around threads and sockets). The latter has much more functionality
as well, but it is not relevant to the benchmarks.