Fix database hostname used by phoenix
Using "127.0.0.1" in the prod config meant that the sed command in
setup.sh, which replaced instances of "localhost" (not "127.0.0.1")
with the correct database hostname read from the DBHOST environment
variable, did not have the desired effect.

"TFB-database" is the correct hostname to use; there's no need to read
the DBHOST environment variable.

This change fixes all the database tests for phoenix.