Fix the ngx_mruby db test so that it prints the world instead of "null"
In the previous version, it was as if the "ret = ..." statement inside
the db.execute block was creating a *new* variable named ret rather than
reassigning the variable that was initialized to nil above, and so the
value of ret seen on the "Nginx.rputs ..." line was always nil.

Strangely (to me), wrapping the whole previous version in a function
definition ("def foo" ... "end") then calling that function ("foo()")
also results in the correct behavior.

I don't know why it behaves this way.  I don't think the implementation
of this test has changed in a long time, and it was fine in previous
rounds.