Remove mysql configuration that was breaking new installations (#3056)
The custom apparmor settings in usr.sbin.mysqld (now deleted) are much
different than the default settings that come with a current mysql
installation; the custom settings just look wrong.  Using the custom
settings causes apparmor-related errors to appear in syslog and causes
a fresh mysql installation to fail.

Setting inndodb_buffer_pool_size to 13G meant that in our local VMs,
mysql would consume all available memory and crash on startup.  Apparently
this is not an issue in the Server Central environment, but 13G seems
unreasonably high.  Isn't the ideal value for that setting proportional
to the amount of data in our database, which is very little?  That seems
to be the case based on what I'm reading here:

  https://dba.stackexchange.com/a/27341