Clear out /tmp after tests, not during, and make it non-optional (#3066)
The "Nuke /tmp" section of the script was problematic for a few
frameworks, since it would clear out /tmp *after* the framework's
server had started but before the tests were run.  If the framework
put files it needed into /tmp upon startup, they'd be deleted right
afterwards.

This causes problems for dancer, for example, who wants to create a
/tmp/perl-dancer.sock file on startup, and when that file gets deleted,
it starts printing a warning message to the log on every request.

The fix is to move the "Nuke /tmp" section so that it runs after the
server is shut down.  Since we're always cleaning up /tmp in that spot
now, there's no reason to support the --clear-tmp argument anymore.