Upgrade activerecord version for php (no framework) to 1.2.0
This should prevent the php test from printing a warning message to the
logs on each request.  The warning messages look like this:

  Server php: 2017/11/13 18:25:57 [error] 7983#0: *44 FastCGI sent in
  stderr: "PHP message: PHP Warning:  Declaration of
  ActiveRecord\DateTime::setTime($hour, $minute, $second = NULL) should
  be compatible with DateTime::setTime($hour, $minute, $second = NULL,
  $microseconds = NULL) in ...

We tried to do this upgrade before, in commit
2430a5978e2d6e40d680f1b6d571a708902efede, but that commit only updated
the composer.json file and not the composer.lock file.  The composer.lock
file was still specifying activerecord 1.1.2, so that was the version
that would get used at runtime.  Removing the lock file means the version
specified in composer.json will get used instead.

We could update the lock file in source control, but since the
composer.json file specifies an exact version for activerecord (rather
than a version range), the lock file doesn't seem to serve any purpose.
Elsewhere in the project, we're inconsistent with whether we have the
composer.lock file in source control or not.