Don't log "connection is closed" and "broken pipe" errors in grizzly-jersey
A huge number of these messages are written to the logs during the
plaintext test.  Broken connections aren't great, but they aren't
necessarily the framework's fault (wrk might be terminating the
connection), and it doesn't help anyone to write the same IOException
stack trace in the log over and over for ~120 MB worth of messages.

With this change, we don't bother writing those messages to the log
anymore.