Fix the multi query database test for blade
It was giving the toolset the wrong URL for the multi query test, so
that the toolset was hitting its single query endpoint instead and
always querying for and responding with one world instead of a variable
number.  That issue is now fixed and the correct endpoint is used.

Surprisingly, our verification step doesn't catch this kind of issue,
and it was satisfied by single-world output.

When I fixed that issue in benchmark_config.json and ran the test again,
it became apparent that the multi query endpoint wasn't handling invalid
parameters correctly.  For example, ?queries=foo would lead to an
uncaught NumberFormatException instead of defaulting to a query count of
1 as specified in the test requirements:

  "If the parameter is missing, is not an integer, or is an integer less
   than 1, the value should be interpreted as 1"

That issue is now fixed as well.